Addition of HiddenKV for hiding object state
Useful for building ETL data mappers (as part of data warehousing), HiddenKV
(and the only current accessor method, HiddenGetStringAlways) allows one to
shuffle object state from the primary KV into an available
map[string]interface{} that lives alongside on the dyndao/object.

HiddenKV is not allocated by default, unlike the other maps in the
dyndao/object Type.

MakeHiddenKVIfNeeded() is available to simplify map allocation when a HiddenKV
becomes necessary.

Currently, it is complex and proprietary to explain the exact situations in
which this can be useful.